migrate to OpenTelemetry, and unify frontend and backend
