--- distributed-tracing/frontend/frontend-configmap.yaml
@@ -1,21 +1,0 @@
-#Copyright 2019 Google LLC
-#Licensed under the Apache License, Version 2.0 (the "License");
-#you may not use this file except in compliance with the License.
-#You may obtain a copy of the License at
-#https://www.apache.org/licenses/LICENSE-2.0
-#Unless required by applicable law or agreed to in writing, software
-#distributed under the License is distributed on an "AS IS" BASIS,
-#WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#See the License for the specific language governing permissions and
-#limitations under the License.
-
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: frontend-configmap
-  namespace: default
-  selfLink: /api/v1/namespaces/default/configmaps/frontend-configmap
-data:
-  PROJECT_ID: "$PROJECT_ID"
-  LOCATION: "us-west1"
-  BACKEND: "http://$BACKEND_IP:8080"
